+++ renderer/index.js
@@ -30,7 +30,6 @@ generateButton.addEventListener("click", (event) => {
   event.preventDefault();
   generateButton.disabled = true;
   const bookCode = document.getElementById("book-select");
-  const createHandouts = document.getElementById("create-handouts");
   const observeAll = document.getElementById("observe-all");
   const messageDiv = document.getElementById("message-div");
   messageDiv.innerHTML = "";
@@ -38,7 +37,6 @@ generateButton.addEventListener("click", (event) => {
   const options = {
     bookCode: bookCode.value,
     observeAll: observeAll.checked,
-    createHandouts: createHandouts.checked,
   };
 
   if (Number.parseInt(bookCode.value) !== 0) {
@@ -116,9 +114,7 @@ window.api.receive("config", (config) => {
       outputLocation.innerHTML = config.data.outputDirEnv;
       window.api.send("books", config);
     }
-    versionSupport.innerHTML = "<p><b>Version:</b> v11+ mode. This version of Adventure Muncher recommends D&D 5e System v3.0.0 or higher.</p>";
-
-    document.getElementById("create-handouts").checked = config.data.createHandouts === true;
+    versionSupport.innerHTML = "<p><b>Version:</b> v11+ mode. This version of Adventure Muncher recommends D&D 5e System v3.0.0 or higher, and DDB Importer v4.1.0 or higher.</p>";
   } else {
     console.warn("No config file!");
     contentLoadMessage.innerHTML = "Config not found";